Moving helpers to another class
