+++ CPKReaderWV/CPKReaderWV.csproj
@@ -50,6 +50,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="helper.cs" />
     <Compile Include="CPKFile.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
